% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspen_qc.R
\name{checkAspenQC}
\alias{checkAspenQC}
\title{Check that all plots have at least one live aspen each time they are visited}
\usage{
checkAspenQC(data = fetchAndWrangleAspen())
}
\arguments{
\item{data}{Aspen dataset to run through QC functions}
}
\description{
Check that all plots have at least one live aspen each time they are visited
}
