% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspen_qc_ucbn.R
\name{checkAspenUCBNQC}
\alias{checkAspenUCBNQC}
\title{Check that all plots have at least one live aspen for each visit}
\usage{
checkAspenUCBNQC(data)
}
\arguments{
\item{data}{Aspen dataset to run through QC functions}
}
\description{
Check that all plots have at least one live aspen for each visit
}
