% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{checkSiteAspen}
\alias{checkSiteAspen}
\title{Check that each site visit has at least one live aspen recorded}
\usage{
checkSiteAspen(aspen_data)
}
\arguments{
\item{aspen_data}{Nested list of aspen dataframes in data package format, as returned by loadAndPackageAspen}
}
\value{
Table containing site visits that do not have at least one live aspen recorded
}
\description{
Check that each site visit has at least one live aspen recorded
}
