% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{checkSiteVisits}
\alias{checkSiteVisits}
\title{Check that all site visits have at least one observation}
\usage{
checkSiteVisits(aspen_data)
}
\arguments{
\item{aspen_data}{Nested list of aspen dataframes in data package format, as returned by loadAndPackageAspen}
}
\value{
Table containing site visits that do not have at least one observation
}
\description{
Check that all site visits have at least one observation
}
