% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchAGOLToken}
\alias{fetchAGOLToken}
\title{Get AGOL authentication token}
\usage{
fetchAGOLToken(
  agol_username,
  agol_password = keyring::key_get(service = "AGOL", username = agol_username),
  root = "nps.maps.arcgis.com",
  referer = "https://irma.nps.gov"
)
}
\arguments{
\item{agol_username}{AGOL headless account username}

\item{agol_password}{AGOL headless account password (do not hard code this into your scripts!)}

\item{root}{NPS users should keep the default. See \url{https://developers.arcgis.com/rest/users-groups-and-items/root.htm} for more information.}

\item{referer}{NPS users should keep the default. See \url{https://developers.arcgis.com/rest/users-groups-and-items/generate-token.htm} for more information.}
}
\value{
An AGOL authentication token
}
\description{
Treat this token as you would a password: don't hard-code it in your scripts or save it to a file. It will expire after 60 minutes.
}
