% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchAllRecords}
\alias{fetchAllRecords}
\title{Fetch tabular data from AGOL}
\usage{
fetchAllRecords(
  data_path,
  layer_number,
  token,
  geometry = FALSE,
  where = "1=1",
  outFields = "*"
)
}
\arguments{
\item{data_path}{Feature service URL}

\item{layer_number}{Layer number}

\item{token}{Authentication token (not needed for public layers)}

\item{geometry}{Include spatial data columns? Works with points, not tested with other geometry types}

\item{where}{Query clause specifying a subset of rows (optional; defaults to all rows). See AGOL REST API documentation.}

\item{outFields}{String indicating which fields to return (optional; defaults to all fields). See AGOL REST API documentation.}
}
\value{
A tibble
}
\description{
Retrieves tabular data from AGOL layers and tables, even when number of rows exceeds maximum record count.
}
