% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchAndWrangleAspen}
\alias{fetchAndWrangleAspen}
\title{Fetch feature service info from AGOL}
\usage{
fetchAndWrangleAspen(
  aspen_url =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_Aspen_Test_Visit_NonSpatial_gdb/FeatureServer",
  site_url =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/AspenSites2/FeatureServer",
  agol_username = "mojn_data"
)
}
\arguments{
\item{url}{Feature service URL}

\item{token}{Authentication token (not needed for public layers)}
}
\value{
A list
}
\description{
Retrieves metadata from AGOL layers and tables.
}
