% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchAndWrangleUCBNAspen}
\alias{fetchAndWrangleUCBNAspen}
\title{Fetch aspen data from AGOL and do preliminary data wrangling}
\usage{
fetchAndWrangleUCBNAspen(
  aspen_url =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/service_4d6343e9204142928351c52c6f1362c5/FeatureServer",
  site_url = NULL,
  agol_username = "mojn_data"
)
}
\arguments{
\item{aspen_url}{URL to main AGOL aspen database}

\item{site_url}{URL to AGOL database for aspen sites (if applicable)}

\item{agol_username}{Authentication token (not needed for public layers)}
}
\value{
A list of data frames and metadata
}
\description{
Fetch aspen data from AGOL and do preliminary data wrangling
}
