% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchMetadata}
\alias{fetchMetadata}
\title{Fetch metadata from AGOL}
\usage{
fetchMetadata(url, token, layer_number)
}
\arguments{
\item{url}{Feature service URL}

\item{token}{Authentication token (not needed for public layers)}

\item{layer_number}{Optional layer ID}
}
\value{
A list
}
\description{
Retrieves metadata from AGOL layers and tables.
}
