% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadAndPackageAspen}
\alias{loadAndPackageAspen}
\title{Load from AGOL, wrangle and package data collected by the Mojave Desert Network and the Upper Columbia Basin Network for the shared Aspen Monitoring Protocol}
\usage{
loadAndPackageAspen(
  network,
  agol_username = "mojn_data",
  agol_password = keyring::key_get(service = "AGOL", username = agol_username)
)
}
\arguments{
\item{network}{String of the network code for which aspen data you wish to retrieve. MOJN or UCBN are accepted values.}

\item{agol_username}{AGOL headless account username}

\item{agol_password}{AGOL headless account password (do not hard code this into your scripts!)}
}
\value{
A list of aspen data frames in data package format
}
\description{
Load from AGOL, wrangle and package data collected by the Mojave Desert Network and the Upper Columbia Basin Network for the shared Aspen Monitoring Protocol
}
