% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadAndWrangleUCBNAspen}
\alias{loadAndWrangleUCBNAspen}
\title{Fetch, wrangle, and package UCBN aspen data from AGOL}
\usage{
loadAndWrangleUCBNAspen(
  aspen_url =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/survey123_11607b1ef8eb4a3397b6686acf61b9d2_results/FeatureServer",
  site_url =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/UCBN_Aspen_Locations_pt_20210430/FeatureServer",
  agol_username = "mojn_data",
  agol_password = keyring::key_get(service = "AGOL", username = agol_username)
)
}
\arguments{
\item{aspen_url}{URL to MOJN_Aspen_Database on AGOL}

\item{site_url}{URL to MOJN_Aspen_Sites_Master on AGOL}

\item{agol_username}{AGOL headless account username}

\item{agol_password}{AGOL headless account password (do not hard code this into your scripts!)}
}
\value{
A list of aspen data frames in data package format
}
\description{
Fetch, wrangle, and package UCBN aspen data from AGOL
}
