% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspen_qc_ucbn.R
\name{missingTreeUCBNQC}
\alias{missingTreeUCBNQC}
\title{Check that there is at least one tree for every observation entry and flag entries with more than 250 trees}
\usage{
missingTreeUCBNQC(data)
}
\arguments{
\item{data}{Aspen dataset to run through QC functions}
}
\description{
Check that there is at least one tree for every observation entry and flag entries with more than 250 trees
}
