% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{packageAspen}
\alias{packageAspen}
\title{Format wrangled aspen data for data package publication}
\usage{
packageAspen(aspen_data)
}
\arguments{
\item{aspen_data}{List of tabular wrangled aspen data}
}
\value{
aspen_data with data package formatting
}
\description{
Format wrangled aspen data for data package publication
}
