% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspen_qc.R
\name{treeChangeQC}
\alias{treeChangeQC}
\title{Flag large changes in number of tree between years}
\usage{
treeChangeQC(data = fetchAndWrangleAspen())
}
\arguments{
\item{data}{Aspen dataset to run through QC functions}
}
\description{
Flag large changes in number of tree between years
}
