% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspen_qc_ucbn.R
\name{treeChangeUCBNQC}
\alias{treeChangeUCBNQC}
\title{Flag large changes in number of tree between years}
\usage{
treeChangeUCBNQC(data)
}
\arguments{
\item{data}{Aspen dataset to run through QC functions}
}
\description{
Flag large changes in number of tree between years
}
