% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspen_qc.R
\name{treeSpeciesQC}
\alias{treeSpeciesQC}
\title{Check that no entries in the observation table have a SpeciesCode that is NA or missing}
\usage{
treeSpeciesQC(data = fetchAndWrangleAspen())
}
\description{
Check that no entries in the observation table have a SpeciesCode that is NA or missing
}
