% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspen_qc_ucbn.R
\name{treeSpeciesUCBNQC}
\alias{treeSpeciesUCBNQC}
\title{Check that no entries in the observation table have a SpeciesCode that is NA or missing}
\usage{
treeSpeciesUCBNQC(data)
}
\arguments{
\item{data}{Aspen dataset to run through QC functions}
}
\description{
Check that no entries in the observation table have a SpeciesCode that is NA or missing
}
