% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspen_qc_ucbn.R
\name{unknownSpeciesUCBNQC}
\alias{unknownSpeciesUCBNQC}
\title{Check that unknown species entries do not have any live trees}
\usage{
unknownSpeciesUCBNQC(data)
}
\arguments{
\item{data}{Aspen dataset to run through QC functions}
}
\description{
Check that unknown species entries do not have any live trees
}
