% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleMOJNObservations}
\alias{wrangleMOJNObservations}
\title{Wrangle observations table}
\usage{
wrangleMOJNObservations(raw_data)
}
\arguments{
\item{raw_data}{Data returned by fetchagol::fetchRawData on MOJN_Aspen_Database}
}
\value{
raw_data with updated observations table
}
\description{
Wrangle observations table
}
