% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleMOJNPests}
\alias{wrangleMOJNPests}
\title{Wrangle pests table}
\usage{
wrangleMOJNPests(raw_data)
}
\arguments{
\item{raw_data}{Data returned by fetchagol::fetchRawData on MOJN_Aspen_Database}
}
\value{
raw_data with updated pests table
}
\description{
Wrangle pests table
}
