% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleMOJNSites}
\alias{wrangleMOJNSites}
\title{Wrangle MOJN sites table}
\usage{
wrangleMOJNSites(raw_data)
}
\arguments{
\item{raw_data}{List of tabular data and metadata returned by using \code{fetchagol::fetchRawData} on MOJN_Aspen_Sites_Master}
}
\value{
raw_data with updated sites table
}
\description{
Wrangle MOJN sites table
}
