% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wranglePests}
\alias{wranglePests}
\title{Wrangle pests table}
\usage{
wranglePests(raw_data)
}
\arguments{
\item{raw_data}{Object returned by fetchagol::fetchRawData on MOJN_Aspen_Database}
}
\value{
raw_data with updated pests table
}
\description{
Wrangle pests table
}
