% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleSites}
\alias{wrangleSites}
\title{Wrangle optional sites table}
\usage{
wrangleSites(raw_data)
}
\arguments{
\item{raw_data}{Object returned by fetchagol::fetchRawData on MOJN_Aspen_Sites_Master}
}
\value{
raw_site_data with updated sites table and metadata fields
}
\description{
Wrangle optional sites table
}
