% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleSiteVisit}
\alias{wrangleSiteVisit}
\title{Wrangle site visit table}
\usage{
wrangleSiteVisit(raw_data)
}
\arguments{
\item{raw_data}{Data returned by fetchagol::fetchRawData on MOJN_Aspen_Database}
}
\value{
raw_data with updated site visit table
}
\description{
Wrangle site visit table
}
