% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleUCBNObservations}
\alias{wrangleUCBNObservations}
\title{Wrangle UCBN observations table}
\usage{
wrangleUCBNObservations(raw_data)
}
\arguments{
\item{raw_data}{Data table returned by fetchagol::fetchRawData on 'UCBN Aspen Survey v1'}
}
\value{
raw_data with updated sites table
}
\description{
Wrangle UCBN observations table
}
