% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleUCBNPests}
\alias{wrangleUCBNPests}
\title{Wrangle UCBN pests table}
\usage{
wrangleUCBNPests(raw_data)
}
\arguments{
\item{raw_data}{Data table returned by fetchagol::fetchRawData on 'UCBN Aspen Survey v1'}
}
\value{
raw_data with updated sites table
}
\description{
Wrangle UCBN pests table
}
