% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeAspen}
\alias{writeAspen}
\title{Write aspen data to CSV}
\usage{
writeAspen(
  all_data,
  data_dir = here::here("data", "final"),
  dictionary_dir = here::here("data", "dictionary"),
  dictionary_filenames = c(tables = "data_dictionary_tables.txt", attributes =
    "data_dictionary_attributes.txt", categories = "data_dictionary_categories.txt"),
  verbose = FALSE,
  removeColumns = TRUE,
  cols_to_remove = c("Editor", "Creator")
)
}
\arguments{
\item{all_data}{Output of \code{fetchRawData()}}

\item{data_dir}{Folder to store data csv's in}

\item{dictionary_dir}{Folder to store data dictionaries in}

\item{dictionary_filenames}{Named list with names \code{c("tables", "attributes", "categories")} indicating what to name the tables, attributes, and categories data dictionaries. You are encouraged to keep the default names unless you have a good reason to change them.}

\item{verbose}{Output feedback to console?}

\item{removeColumns}{Should columns be removed?}

\item{cols_to_remove}{Columns that should be removed, \code{c("Editor", "Creator")} is the default because they can contain personally identifiable information}
}
\description{
Write aspen data to CSV
}
