% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspen_qc.R
\name{checkAspenQC}
\alias{checkAspenQC}
\title{Check that all of the plots have at least one aspen each time they are visited}
\usage{
checkAspenQC(data = fetchAndWrangleAspen())
}
\description{
Check that all of the plots have at least one aspen each time they are visited
}
