% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{checkDuplicateTrees}
\alias{checkDuplicateTrees}
\title{Check for duplicate tree records in the Observations table}
\usage{
checkDuplicateTrees(aspen_data)
}
\arguments{
\item{aspen_data}{List of aspen dataframes in data package format, as returned by loadAndPackageAspen}
}
\value{
Table containing duplicate tree records in the Observations table
}
\description{
Check for duplicate tree records in the Observations table
}
