% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{checkTreeCount}
\alias{checkTreeCount}
\title{Check for tree records in the Observations table that have a count of zero or a count greater than 250}
\usage{
checkTreeCount(aspen_data)
}
\arguments{
\item{aspen_data}{List of aspen dataframes in data package format, as returned by loadAndPackageAspen}
}
\value{
Table containing tree records that have a count of zero or a count greater than 250
}
\description{
Check for tree records in the Observations table that have a count of zero or a count greater than 250
}
