% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{checkTreeID}
\alias{checkTreeID}
\title{Check that all tree species have been identified}
\usage{
checkTreeID(aspen_data)
}
\arguments{
\item{aspen_data}{List of aspen dataframes in data package format, as returned by loadAndPackageAspen}
}
\value{
Table containing rows that are missing either a species code, verbatim identification, or scientific name
}
\description{
Check that all tree species have been identified
}
