% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_network}
\alias{get_network}
\title{Get network based on content of data}
\usage{
get_network(data, raw = FALSE)
}
\arguments{
\item{data}{list of aspen dataframes}
}
\value{
Network code
}
\description{
Get network based on content of data
}
