% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadAndWrangleMOJNAspen}
\alias{loadAndWrangleMOJNAspen}
\title{Fetch aspen data from AGOL and do preliminary data wrangling}
\usage{
loadAndWrangleMOJNAspen(
  aspen_url =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_Aspen_Test_Visit_NonSpatial_gdb/FeatureServer",
  site_url =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/AspenSites2/FeatureServer",
  agol_username = "mojn_data"
)
}
\arguments{
\item{aspen_url}{URL to main AGOL aspen database}

\item{site_url}{URL to AGOL database for aspen sites (if applicable)}

\item{agol_username}{Authentication token (not needed for public layers)}
}
\value{
A list of data frames and metadata
}
\description{
Fetch aspen data from AGOL and do preliminary data wrangling
}
