% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{packageMOJNAspen}
\alias{packageMOJNAspen}
\title{Format aspen data for data package publication}
\usage{
packageMOJNAspen(aspen_data)
}
\arguments{
\item{aspen_data}{The wrangled aspen data}
}
\value{
aspen_data with data package formatting
}
\description{
Format aspen data for data package publication
}
