% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{packageUCBNAspen}
\alias{packageUCBNAspen}
\title{Format aspen data for data package publication}
\usage{
packageUCBNAspen(aspen_data)
}
\arguments{
\item{aspen_data}{The wrangled UCBN aspen data}
}
\value{
aspen_data with data package formatting
}
\description{
Format aspen data for data package publication
}
