% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleDisturbances}
\alias{wrangleDisturbances}
\title{Wrangle disturbances table}
\usage{
wrangleDisturbances(raw_data)
}
\arguments{
\item{raw_data}{Object returned by fetchagol::fetchRawData on MOJN_Aspen_Database}
}
\value{
raw_data with updated disturbances table and metadata fields
}
\description{
Wrangle disturbances table
}
