% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleDisturbances}
\alias{wrangleDisturbances}
\title{#' Wrangle aspen disturbances table}
\usage{
wrangleDisturbances(raw_data)
}
\arguments{
\item{raw_data}{List of tabular data and metadata containing the results returned by using \code{fetchagol::fetchRawData} on MOJN_Aspen_Sites_Master and MOJN_Aspen_Database or on UCBN_Aspen_Locations_pt_20210430 and 'UCBN Aspen Survey v1'}
}
\value{
raw_data with updated sites table
}
\description{
#' Wrangle aspen disturbances table
}
