% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleMOJNDisturbances}
\alias{wrangleMOJNDisturbances}
\title{Wrangle disturbances table}
\usage{
wrangleMOJNDisturbances(raw_data)
}
\arguments{
\item{raw_data}{Data returned by fetchagol::fetchRawData on MOJN_Aspen_Database}
}
\value{
raw_data with updated disturbances table
}
\description{
Wrangle disturbances table
}
