% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleMOJNSites}
\alias{wrangleMOJNSites}
\title{Wrangle sites table}
\usage{
wrangleMOJNSites(raw_data)
}
\arguments{
\item{raw_data}{Data table returned by fetchagol::fetchRawData on MOJN_Aspen_Sites_Master}
}
\value{
raw_data with updated sites table
}
\description{
Wrangle sites table
}
