% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleUCBNDisturbances}
\alias{wrangleUCBNDisturbances}
\title{Wrangle UCBN disturbances table}
\usage{
wrangleUCBNDisturbances(raw_data)
}
\arguments{
\item{raw_data}{List of tabular data and metadata containing the results returned by using \code{fetchagol::fetchRawData} on UCBN_Aspen_Locations_pt_20210430 and 'UCBN Aspen Survey v1'}
}
\value{
raw_data with updated sites table
}
\description{
Wrangle UCBN disturbances table
}
