% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangleUCBNSites}
\alias{wrangleUCBNSites}
\title{Wrangle UCBN sites table}
\usage{
wrangleUCBNSites(raw_data)
}
\arguments{
\item{raw_data}{List of tabular data and metadata returned by using \code{fetchagol::fetchRawData} on UCBN_Aspen_Locations_pt_20210430}
}
\value{
raw_data with updated sites table
}
\description{
Wrangle UCBN sites table
}
